split carzam.py and app.py and updated the README and .gitignore, still haven't implemented file storage clean up.
