+++ README.md
@@ -1,2 +1,13 @@
 # carzam
 The CS467 Capstone Project at Oregon State University in the Spring of 2021. Carzam is a web application that leverages an AI to identify vehicles, their make and model in a submitted picture.
+
+ To test:
+ Ideally, you should setup a virtual environment before running this:
+ python3 -m venv venv
+ source ./venv/bin/activate
+
+ Install dependencies by doing
+ pip3 install -r requirements.txt
+
+ Then run by doing
+ python3 app.py
